Feature/snap fsm update (#628)
