+++ core/api/storage_miner/storage_api.hpp
@@ -259,6 +259,11 @@ namespace fc::api {
                const CallId &,
                Proof,
                const boost::optional<CallError> &)
+    API_METHOD(SectorMarkForUpgrade,
+               jwt::kAdminPermission,
+               void,
+               SectorNumber,
+               bool /* snap_deal */)
 
     API_METHOD(WorkerConnect, jwt::kAdminPermission, void, const std::string &);
   };
@@ -314,6 +319,7 @@ namespace fc::api {
     f(a.ReturnReplicaUpdate);
     f(a.ReturnProveReplicaUpdate1);
     f(a.ReturnProveReplicaUpdate2);
+    f(a.SectorMarkForUpgrade);
     f(a.WorkerConnect);
   }
 }  // namespace fc::api
