+++ core/miner/storage_fsm/sealing.hpp
@@ -59,6 +59,8 @@ namespace fc::mining {
 
     virtual outcome::result<void> markForUpgrade(SectorNumber id) = 0;
 
+    virtual outcome::result<void> markForSnapUpgrade(SectorNumber id) = 0;
+
     virtual bool isMarkedForUpgrade(SectorNumber id) = 0;
 
     virtual outcome::result<void> startPacking(SectorNumber id) = 0;
@@ -79,6 +81,8 @@ namespace fc::mining {
     kFailSubmit,
     kSectorAllocatedError,
     kNotPublishedDeal,
+    kCannotMarkInactiveSector,
+    kSectorExpirationError,
   };
 }  // namespace fc::mining
